Refactor document data to repository and models
Replaces DrawDocumentData with a new DrawDocumentRepository and DrawDocumentModels, migrating all usages to the new structure. Updates all references from DrawDocumentData.Page and related types to the new Page and Stroke models. Improves separation of concerns by moving data models to DrawDocumentModels and repository logic to DrawDocumentRepository, and updates ViewModel and UI code accordingly.
