+++ app/src/main/java/com/studiomath/pencilnotes/document/compose/UsageExample.kt
@@ -4,12 +4,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.studiomath.pencilnotes.document.compose.lazyDocument.LazyDocumentViewer
 import com.studiomath.pencilnotes.document.compose.lazyDocument.items
-import com.studiomath.pencilnotes.document.page.DrawDocumentData
+import com.studiomath.pencilnotes.document.page.Page
 import com.studiomath.pencilnotes.document.page.PageMaker
 
 @Composable
 fun LazyDocumentViewerExample(
-    pages: List<DrawDocumentData.Page>,
+    pages: List<Page>,
     pageMaker: PageMaker
 ) {
     LazyDocumentViewer {
